+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog of lizard-map
 3.4 (unreleased)
 ----------------
 
+- Added layout_extra_from_request to AdapterMixin.
+
 - Updated adapter views for image and values: you can now provide start and
   end dates in the url parameters.